+++ R/zzz.R
@@ -24,8 +24,6 @@ mlr3fda_pipeop_tags = "fda"
 
 named_union = function(x, y) set_names(union(x, y), union(names(x), names(y)))
 
-# metainf must be manually added in the register_mlr3pipelines function
-# Because the value is substituted, we cannot pass it through this function
 register_po = function(name, constructor) {
   if (name %chin% names(mlr3fda_pipeops)) {
     stopf("pipeop %s registered twice", name)
@@ -54,7 +52,7 @@ register_mlr3 = function(...) {
 register_mlr3pipelines = function(...) {
   mlr_reflections = utils::getFromNamespace("mlr_reflections", ns = "mlr3")
   mlr_pipeops = utils::getFromNamespace("mlr_pipeops", ns = "mlr3pipelines")
-  iwalk(as.list(mlr3fda_pipeops), \(value, name) mlr_pipeops$add(name, value$constructor, value$metainf))
+  iwalk(as.list(mlr3fda_pipeops), \(value, name) mlr_pipeops$add(name, value$constructor))
   mlr_reflections$pipeops$valid_tags = union(mlr_reflections$pipeops$valid_tags, mlr3fda_pipeop_tags)
 }